// angular/projects/account/public/config/src/enums/route-names.ts
export const enum eAccountRouteNames {
  Account = 'AbpAccount::Menu:Account',
  Login = 'AbpAccount::Login',
  Register = 'AbpAccount::Register',
  EmailConfirmation = 'AbpAccount::EmailConfirmation',
  LinkLogged = 'AbpAccount::LinkLogged',
  ForgotPassword = 'AbpAccount::ForgotPassword',
  ResetPassword = 'AbpAccount::ResetPassword',
  SendSecurityCode = 'AbpAccount::TwoFactorVerification',
  ManageProfile = 'AbpAccount::MyAccount',
  MySecurityLogs = 'AbpAccount::MySecurityLogs',
}